lou-media-stack/ai-services/storage-optimization/storage_optimization.py
#!/usr/bin/env python3
# AI-Powered Storage Management Service

import os
import time
import logging
from datetime import datetime
from typing import Dict, List

import numpy as np
import pandas as pd
from flask import Flask, request, jsonify
import psutil
import joblib
import psycopg2
from watchdog.observers import Observer
from watchdog.events import FileSystemEventHandler

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class StorageManagementAI:
    def __init__(self):
        self.model_path = '/models/storage_model.joblib'  # Placeholder for serialized model
        self.trend_data_path = '/data/storage_trend_data.csv'  # Placeholder for trend data
        
        self.load_models()
        self.db_connection = psycopg2.connect(
            dbname=os.getenv('POSTGRES_DB', 'mydb'),
            user=os.getenv('POSTGRES_USER', 'myuser'),
            password=os.getenv('POSTGRES_PASSWORD', 'mypassword'),
            host=os.getenv('POSTGRES_HOST', 'localhost'),
            port=os.getenv('POSTGRES_PORT', 5432)
        )

    def load_models(self) -> None:
        """Load serialized models"""
        try:
            if os.path.exists(self.model_path):
                self.storage_model = joblib.load(self.model_path)
                logger.info("Loaded storage model")
            else:
                self.storage_model = None
        except Exception as e:
            logger.error(f"Error loading models: {e}")

    def predict_cleanup(self) -> List[Dict]:
        """Predict files for potential cleanup based on trends"""
        try:
            # Monitor file system changes
            handler = FileSystemEventHandler()
            observer = Observer()
            observer.schedule(handler, path='/media', recursive=True)
            observer.start()

            # Watch for changes and decide what files can be cleaned
            cleanup_candidates = self.analyze_storage_trends()
            return cleanup_candidates
        except Exception as e:
            logger.error(f"Error in predicting cleanup: {e}")
            return []

    def analyze_storage_trends(self) -> List[Dict]:
        """Analyze storage trends to determine cleanup targets"""
        # Example trend analysis logic
        if os.path.exists(self.trend_data_path):
            trend_data = pd.read_csv(self.trend_data_path)
            logger.info("Analyzed storage trends")
            # Example: Predict using trend data
            return [{'file': '/media/old_file.mov', 'reason': 'duplicate'},
                    {'file': '/media/unused_file.avi', 'reason': 'unused'}]
        else:
            return []

    def optimize_storage_usage(self, usage_metrics: Dict) -> Dict:
        """Optimize storage usage and manage files accordingly"""
        # Use metrics to determine if storage optimization is needed
        total, used, free = psutil.disk_usage('/')
        usage_percentage = (used / total) * 100
        logger.info(f"Current storage usage: {usage_percentage:.2f}%")

        # Simulate optimization actions
        if usage_percentage > 80:
            # Heavy optimization required
            return {'action': 'reduce_quality', 'files': ['/media/high_res_movie.mp4']}  # Example
        elif usage_percentage > 50:
            # Moderate optimization
            return {'action': 'remove_duplicates', 'files': ['/media/duplicate_song.mp3']}
        else:
            return {'action': 'no_optimization_needed', 'message': 'Storage usage is optimal'}

# Flask API
app = Flask(__name__)
storage_manager = StorageManagementAI()

@app.route('/health', methods=['GET'])
def health_check():
    """Health check endpoint"""
    return jsonify({
        'status': 'healthy',
        'service': 'AI Storage Optimizer',
        'version': '1.0.0',
        'timestamp': datetime.utcnow().isoformat()
    })

@app.route('/cleanup', methods=['POST'])
def cleanup_files():
    """Endpoint for cleanup prediction"""
    try:
        cleanup_candidates = storage_manager.predict_cleanup()
        return jsonify(cleanup_candidates)
    except Exception as e:
        logger.error(f"Error in cleanup endpoint: {e}")
        return jsonify({ 'error': str(e) }), 500

@app.route('/optimize', methods=['POST'])
def optimize_storage():
    """Optimize and manage storage usage"""
    try:
        data = request.json or {}
        optimized_action = storage_manager.optimize_storage_usage(data)
        return jsonify(optimized_action)
    except Exception as e:
        logger.error(f"Error in storage optimization endpoint: {e}")
        return jsonify({ 'error': str(e) }), 500

if __name__ == '__main__':
    app.run(host='0.0.0.0', port=8779, debug=False)
